Adding null check before deleting subpopulation.

+++ src/test/java/org/sagebionetworks/bridge/sdk/integration/SubpopulationTest.java
@@ -34,7 +34,9 @@ public class SubpopulationTest {
     
     @After
     public void after() {
-        admin.getSession().getAdminClient().deleteSubpopulationPermanently(guid);
+        if (guid != null) {
+            admin.getSession().getAdminClient().deleteSubpopulationPermanently(guid);    
+        }
         developer.signOutAndDeleteUser();
     }
